refactoring AirplayBonjour to see if there is any difference in shutting down aka de-announcing
